+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 * Disables iPad sleep timer.
 * Moves from Balanced to Stripe for credit card processing.
 * Allows registration with email addresses containing plus signs.
+* New and improved bidder registration details screen.
 
 ## 3.4.0 April 13 2015
 * Fixes problems entering in phone numbers.